[Scheduler] Split PatchSchedulerConfig out of PatchScheduler.hpp
Solver configs only need the split/merge load thresholds, but including
PatchScheduler.hpp pulled json, trees, MPI, and SYCL into SPH/GSPH/Ramses
config TUs. Move the struct to a tiny header with json_fwd and out-of-line
to_json/from_json.
