+++ src/shamrock/include/shamrock/scheduler/PatchScheduler.hpp
@@ -23,11 +23,11 @@
 #include "shambase/DistributedData.hpp"
 #include "shambase/stacktrace.hpp"
 #include "shambase/time.hpp"
+#include "nlohmann/json_fwd.hpp"
 #include "shamalgs/collective/distributedDataComm.hpp"
 #include "shamrock/legacy/patch/utility/patch_field.hpp"
 #include "shamrock/solvergraph/PatchDataLayerRefs.hpp"
 #include "shamsolvergraph/node/NodeSetEdge.hpp"
-#include <nlohmann/json.hpp>
 #include <unordered_set>
 #include <fstream>
 #include <functional>
@@ -45,39 +45,11 @@
 #include "shamrock/patch/PatchDataLayerLayout.hpp"
 #include "shamrock/patch/PatchField.hpp"
 #include "shamrock/scheduler/HilbertLoadBalance.hpp"
+#include "shamrock/scheduler/PatchSchedulerConfig.hpp"
 #include "shamrock/scheduler/PatchTree.hpp"
 #include "shamrock/scheduler/SchedulerPatchData.hpp"
 #include "shamsolvergraph/SolverGraphSerializable.hpp"
 #include "shamsys/legacy/sycl_handler.hpp"
-
-struct PatchSchedulerConfig {
-    u64 split_load_value = 0_u64;
-    u64 merge_load_value = 0_u64;
-};
-
-/**
- * @brief Converts a PatchSchedulerConfig object to a JSON object.
- *
- * @param j The JSON object to be populated.
- * @param p The PatchSchedulerConfig object to be converted.
- */
-inline void to_json(nlohmann::json &j, const PatchSchedulerConfig &p) {
-    j = nlohmann::json{
-        {"split_load_value", p.split_load_value},
-        {"merge_load_value", p.merge_load_value},
-    };
-}
-
-/**
- * @brief Deserializes a PatchSchedulerConfig object from a JSON object.
- *
- * @param j The JSON object to deserialize from.
- * @param p The PatchSchedulerConfig object to populate.
- */
-inline void from_json(const nlohmann::json &j, PatchSchedulerConfig &p) {
-    j.at("split_load_value").get_to<u64>(p.split_load_value);
-    j.at("merge_load_value").get_to<u64>(p.merge_load_value);
-}
 
 /**
  * @brief The MPI scheduler
